Refactor: Improve environment details and fix mock output parsing
