+++ tests/unit/test_operator.py
@@ -1,5 +1,3 @@
-import subprocess
-from datetime import datetime
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
@@ -291,86 +289,3 @@ async def test_operator_print_hello_world(cli_operator):
     assert last_message_content.response == "I have printed 'Hello World' to the console."
     assert last_message_content.code == ""
     assert last_message_content.action == ActionType.DONE
-
-
-def test_get_environment_details(cli_operator, monkeypatch, tmp_path):
-    """Test that get_environment_details returns correct environment information."""
-    # Mock directory indexing
-    mock_index = {
-        ".": [
-            ("test.py", "code", 1024),
-            ("doc.txt", "doc", 500),
-            ("image.png", "image", 2048576),
-            ("other.bin", "other", 750),
-        ]
-    }
-    monkeypatch.setattr("local_operator.operator.index_current_directory", lambda: mock_index)
-
-    # Mock git status
-    def mock_check_output(*args, **kwargs):
-        return b"On branch main\nnothing to commit, working tree clean\n"
-
-    monkeypatch.setattr("subprocess.check_output", mock_check_output)
-
-    # Mock current working directory and datetime
-    monkeypatch.setattr("os.getcwd", lambda: str(tmp_path))
-    fixed_datetime = datetime(2024, 1, 1, 12, 0, 0)
-    monkeypatch.setattr(
-        "local_operator.operator.datetime",
-        type("MockDateTime", (), {"now": lambda: fixed_datetime}),
-    )
-
-    # Get environment details
-    env_details = cli_operator.get_environment_details()
-
-    # Verify output contains expected information
-    assert str(tmp_path) in env_details
-    assert "2024-01-01 12:00:00" in env_details
-    assert "On branch main" in env_details
-    assert "nothing to commit, working tree clean" in env_details
-
-    # Verify directory tree formatting
-    assert "📁 ./" in env_details
-    assert "📄 test.py (code, 1.0KB)" in env_details
-    assert "📝 doc.txt (doc, 500B)" in env_details
-    assert "🖼️ image.png (image, 2.0MB)" in env_details
-    assert "📎 other.bin (other, 750B)" in env_details
-
-
-def test_get_environment_details_no_git(cli_operator, monkeypatch, tmp_path):
-    """Test get_environment_details when not in a git repository."""
-    # Mock directory indexing with empty directory
-    monkeypatch.setattr("local_operator.operator.index_current_directory", lambda: {})
-
-    # Mock git branch check to fail
-    def mock_check_output(*args, **kwargs):
-        raise subprocess.CalledProcessError(128, "git")
-
-    monkeypatch.setattr("subprocess.check_output", mock_check_output)
-
-    # Mock current working directory
-    monkeypatch.setattr("os.getcwd", lambda: str(tmp_path))
-
-    env_details = cli_operator.get_environment_details()
-
-    assert "Not a git repository" in env_details
-
-
-def test_get_environment_details_large_directory(cli_operator, monkeypatch):
-    """Test get_environment_details handles large directories correctly."""
-    # Create mock directory with >300 files
-    mock_files = [("file{}.txt".format(i), "doc", 100) for i in range(1000)]
-    mock_index = {f"dir{i}": mock_files[i * 100 : (i + 1) * 100] for i in range(10)}
-    monkeypatch.setattr("local_operator.operator.index_current_directory", lambda: mock_index)
-
-    # Mock git branch
-    def mock_check_output(*args, **kwargs):
-        return b"main\n"
-
-    monkeypatch.setattr("subprocess.check_output", mock_check_output)
-
-    env_details = cli_operator.get_environment_details()
-
-    # Verify file count limits are enforced
-    assert "... and more files" in env_details
-    assert "... and 70 more files" in env_details
